+++ airflow-core/src/airflow/ui/openapi-gen/requests/schemas.gen.ts
@@ -6132,7 +6132,7 @@ export const $TaskResponse = {
             ]
         },
         retry_exponential_backoff: {
-            type: 'boolean',
+            type: 'number',
             title: 'Retry Exponential Backoff'
         },
         priority_weight: {