+++ providers/slack/docs/changelog.rst
@@ -40,11 +40,15 @@ Misc
 
 * ``Convert all airflow distributions to be compliant with ASF requirements (#58138)``
 * ``Migrate slack provider to ''common.compat'' (#57110)``
+* ``Add 'use-next-version' comment in providers that will need rc2 (#58390)``
+* ``Remove SDK reference for NOTSET in Airflow Core (#58258)``
 
 .. Below changes are excluded from the changelog. Move them to
    appropriate section above if needed. Do not delete the lines(!):
    * ``Delete all unnecessary LICENSE Files (#58191)``
    * ``Enable PT006 rule to slack Provider test (#57963)``
+   * ``Updates to release process of providers (#58316)``
+   * ``Update documentation for providers 14 Nov 2025 (#58284)``
 
 9.4.0
 .....